requirements.txt
aniso8601==9.0.1
autopep8==1.5.7
cachelib==0.4.1
cffi==1.15.0
click==8.0.3
cryptography==35.0.0
decorator==5.1.0
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
flask-marshmallow==0.14.0
Flask-MySQL==1.5.2
Flask-RESTful==0.3.8
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.2
idna==3.3
infinity==1.5
intervals==0.9.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
marshmallow==3.14.0
marshmallow-sqlalchemy==0.26.1
mysql-connector==2.2.9
mysql-connector-python==8.0.26
protobuf==3.18.1
pycodestyle==2.8.0
pycparser==2.20
PyJWT==1.7.1
PyMySQL==1.0.2
python-dotenv==0.19.1
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.25
toml==0.10.2
validators==0.18.2
Werkzeug==2.0.2
WTForms==2.3.3
WTForms-Components==0.10.5
